% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_listL.R
\name{add_listL}
\alias{add_listL}
\title{Add column for list lengths to the data}
\usage{
add_listL(obdata)
}
\arguments{
\item{obdata}{Data frame containing species occurrence records with the following columns: Species, Date, Gridref, Year, Week, Month, and optionally covnames (see below) and listL}
}
\description{
Add column for list lengths to the data
}
