% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_occ.R
\name{fit_occ}
\alias{fit_occ}
\title{Function to fit the occupancy models}
\usage{
fit_occ(
  spp,
  obdata,
  occformula = "~North+I(North^2)+East+I(East^2)",
  detformula = "~logLL+SEAS",
  covnames = c("East", "North"),
  years = NULL,
  minyear = NULL,
  maxyear = NULL,
  trendyears = NULL,
  allsites = NULL,
  qu = FALSE,
  qval = NULL,
  outputdir = NULL,
  multistart = FALSE,
  printprogress = FALSE,
  prev_start = NULL,
  engine = engine,
  prev_output = NULL
)
}
\arguments{
\item{spp}{Target species to estimate occupancy for.}

\item{obdata}{Data frame containing species occurrence records with the following columns: Species, Date, Gridref, Year, Week, Month, and optionally covnames (see below) and listL}

\item{occformula}{Formula for occupancy probability}

\item{detformula}{Formula for detection probability}

\item{covnames}{Vector of covariate names in obdata}

\item{minyear}{First year of interest}

\item{maxyear}{Last year of interest}

\item{trendyears}{Vector of start years for trend estimation. If \code{trendyears = NULL} then no trends will be calculated.}

\item{outputdir}{Directory for output files}
}
\value{
A list containing various outputs
}
\description{
Function to fit the occupancy models
}
