% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Function to estimate occupancy trends}
\usage{
fit_trend(
  z,
  startyear = NULL,
  endyear = NULL,
  predictions = FALSE,
  regionalise = NULL
)
}
\arguments{
\item{z}{Data frame output from \code{fit_occ}, containing estimated occupancy index and standard errors.}

\item{startyear}{Optional. First year to calculate the trend from. Default will select the first year of the index.}

\item{endyear}{Optional. Year to calculate the trend up to. Default will select the last year of the index.}

\item{predictions}{Option to include predicted values in the output e.g. for plotting.}
}
\description{
Function to estimate occupancy trends
}
