% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_occ.R
\name{fit_occ}
\alias{fit_occ}
\title{Function to fit the occupancy models}
\usage{
fit_occ(
  spp,
  obdata,
  occformula = "~North+I(North^2)+East+I(East^2)",
  detformula = "~logLL+SEAS",
  covnames = c("East", "North"),
  minyear = NULL,
  maxyear = NULL,
  trendyears = NULL,
  nstart = 1,
  allsites = NULL,
  qval = NULL,
  prev_start = NULL,
  outputdir = NULL,
  printprogress = FALSE,
  engine = engine,
  prev_output = NULL,
  irun = 0
)
}
\arguments{
\item{spp}{Target species to estimate occupancy for.}

\item{obdata}{Data frame containing species occurrence records with the following columns: Species, Date, Gridref, Year, Week, Month, and optionally covnames (see below) and listL}

\item{occformula}{Formula for occupancy probability}

\item{detformula}{Formula for detection probability}

\item{covnames}{Vector of covariate names in obdata}

\item{minyear}{First year of interest}

\item{maxyear}{Last year of interest}

\item{trendyears}{Vector of start years for trend estimation. If \code{trendyears = NULL} then no trends will be calculated.}

\item{nstart}{Number of starting values to run. Default \code{nstart = 3}.}

\item{allsites}{Optional data frame of sites for which the occupancy index will be calculated for.}

\item{qval}{Quantile value to filter records to months where the species was observed. Default \code{qval = 0.025}.}

\item{prev_start}{Provide starting values e.g. based on outputs of a previous run.}

\item{outputdir}{Optional directory to save output files to.}

\item{printprogress}{Print the progress of the run (only available for non-parallel option)}

\item{engine}{Choose the engine used by unmarked.}

\item{prev_output}{Previous output to append.}
}
\value{
A list containing various outputs
}
\description{
Function to fit the occupancy models
}
