% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_occ_ms.R
\name{fit_occ_ms}
\alias{fit_occ_ms}
\title{Wrapper function to fit occupancy models to multiple species.
#' Function to fit the occupancy models}
\usage{
fit_occ_ms(
  ispp,
  obdata,
  occformula = "~North+I(North^2)+East+I(East^2)",
  detformula = "~logLL+SEAS",
  covnames = c("East", "North"),
  parallel = FALSE,
  cpus = NULL,
  minyear = NULL,
  maxyear = NULL,
  trendyears = NULL,
  allsites = NULL,
  qu = FALSE,
  qval = NULL,
  outputdir = NULL,
  multistart = FALSE,
  printprogress = FALSE,
  prev_start = NULL,
  engine = "R"
)
}
\arguments{
\item{ispp}{Vector of target species to model}

\item{obdata}{Data frame containing species occurrence records with the following columns: Species, Date, Gridref, Year, Week, Month, and optionally covnames (see below) and listL}

\item{occformula}{Formula for occupancy probability}

\item{detformula}{Formula for detection probability}

\item{covnames}{Vector of covariate names in obdata}

\item{parallel}{Logical. Whether to run multiple species in parallel. Default is \code{FALSE}.}

\item{cpus}{Optional specification for the number of cpus when \code{parallel = TRUE}. Otherwise chosen based on cores available and the number of species.}

\item{minyear}{First year of interest}

\item{maxyear}{Last year of interest}

\item{trendyears}{Vector of start years for trend estimation. If \code{trendyears = NULL} then no trends will be calculated.}

\item{outputdir}{Directory for output files}
}
\value{
A list containing various outputs
}
\description{
Wrapper function to fit occupancy models to multiple species.
#' Function to fit the occupancy models
}
