% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Function to estimate occupancy trends}
\usage{
fit_trend(startyear, z, endyear = NULL, predictions = FALSE)
}
\arguments{
\item{startyear}{First year to calculate the trend from}

\item{z}{Data frame output from \code{fit_occ}, containing estimated occupancy index and standard errors.}

\item{endyear}{Optional. Year to calculate the trend up to. Default will select the last year of the index.}

\item{predictions}{Option to include predicted values in the output e.g. for plotting.}
}
\description{
Function to estimate occupancy trends
}
